Ya funciona el z-index. Al momento de crear un sprite se le debe asginar el z-index y listo. Ventana Grafica se encarga de dibujarTodo en orden. Los numeros menores van al fondo, los mayores hacia adelante.

+++ trunk/src/vista/Sprite.cpp
@@ -78,7 +78,7 @@ void Sprite::setFotogramaActual(int nroFotograma){
 	this->fotogramaActual = nroFotograma;
 }
 
-void Sprite::setZindex(int z_index){
+void Sprite::setZindex(float z_index){
 	this->zIndex = z_index;
 }
